check if qemu binary is set

+++ tasks/stop-qemu-box/task.bash
@@ -1,5 +1,12 @@
 qemu_binary=$(config qemu_binary)
 
+if test "$qemu_binary" -eq ""; then
+    echo "qemu_binary is not set";
+    exit 1 
+fi
+
+which $qemu_binary || exit 1
+
 echo "killing by $qemu_binary"
 
 ps aux | grep "$qemu_binary" | grep -v grep | awk '{print "kill -9 " $2}' | sh
